requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
aiosqlite==0.20.0
python-telegram-bot==21.0
openai==1.50.0
pydantic==2.9.0
python-dotenv==1.0.0
APScheduler==3.11.2
